Git pull block for deployments

+++ proxy/flows.py
@@ -1,16 +1,24 @@
 """Reusable flows"""
 
 import os
-from fastapi import HTTPException
 from prefect import flow
+from prefect.blocks.system import Secret
 from prefect_airbyte.flows import run_connection_sync
 from prefect_airbyte import AirbyteConnection
-from prefect_dbt.cli.commands import DbtCoreOperation
+from prefect_dbt.cli.commands import DbtCoreOperation, ShellOperation
 from logger import logger
+
+
+# django prefect block names
+AIRBYTESERVER = "Airbyte Server"
+AIRBYTECONNECTION = "Airbyte Connection"
+SHELLOPERATION = "Shell Operation"
+DBTCORE = "dbt Core Operation"
 
 
 @flow
 def run_airbyte_connection_flow(block_name: str):
+    # pylint: disable=broad-exception-caught
     """Prefect flow to run airbyte connection"""
     try:
         airbyte_connection = AirbyteConnection.load(block_name)
@@ -18,28 +26,32 @@ def run_airbyte_connection_flow(block_name: str):
         logger.info("airbyte connection sync result=")
         logger.info(result)
         return result
-    except Exception as error:
+    except Exception as error:  # skipcq PYL-W0703
+        # pylint: disable=broad-exception-caught
         # logger.exception(error)
         logger.error(str(error))  # "Job <num> failed."
-        raise HTTPException(status_code=400, detail=str(error)) from error
+
+    return None
 
 
 @flow
 def run_dbtcore_flow(block_name: str):
+    # pylint: disable=broad-exception-caught
     """Prefect flow to run dbt"""
     try:
         dbt_op = DbtCoreOperation.load(block_name)
         if os.path.exists(dbt_op.profiles_dir / "profiles.yml"):
             os.unlink(dbt_op.profiles_dir / "profiles.yml")
         return dbt_op.run()
-    except Exception as error:
-        # logger.exception(error)
-        logger.error("FAILED")
-        raise HTTPException(status_code=400, detail=str(error)) from error
+    except Exception as error:  # skipcq PYL-W0703
+        logger.exception(error)
+
+    return None
 
 
 @flow
 def deployment_schedule_flow(airbyte_blocks: list, dbt_blocks: list):
+    # pylint: disable=broad-exception-caught
     """A general flow function that will help us create deployments"""
     # sort the dbt blocks by seq
     dbt_blocks.sort(key=lambda blk: blk["seq"])
@@ -52,17 +64,42 @@ def deployment_schedule_flow(airbyte_blocks: list, dbt_blocks: list):
         airbyte_connection = AirbyteConnection.load(block["blockName"])
         try:
             run_connection_sync(airbyte_connection)
-        except Exception as error:
+        except Exception as error:  # skipcq PYL-W0703
             logger.exception(error)
-            raise HTTPException(status_code=400, detail=str(error)) from error
 
     # run dbt blocks
     for block in dbt_blocks:
-        dbt_op = DbtCoreOperation.load(block["blockName"])
-        if os.path.exists(dbt_op.profiles_dir / "profiles.yml"):
-            os.unlink(dbt_op.profiles_dir / "profiles.yml")
-        try:
-            dbt_op.run()
-        except Exception as error:
-            logger.exception(error)
-            raise HTTPException(status_code=400, detail=str(error)) from error
+        if block["blockType"] == SHELLOPERATION:
+            shell_op = ShellOperation.load(block["blockName"])
+
+            try:
+                # fetch the secret block having the git oauth token-based url to pull code from private repos
+                # the key "secret-git-pull-url-block" will always be present. Value will be empty string if no token was submitted by user
+                secret_block_name = shell_op.env["secret-git-pull-url-block"]
+                git_repo_endpoint = ""
+                if secret_block_name and len(secret_block_name) > 0:
+                    secret_blk = Secret.load(secret_block_name)
+                    git_repo_endpoint = secret_blk.get()
+
+                # update the commands to account for the token
+                commands = shell_op.commands
+                updated_cmds = []
+                for cmd in commands:
+                    updated_cmds.append(f"{cmd} {git_repo_endpoint}")
+                shell_op.commands = updated_cmds
+
+                # run the shell command(s)
+                shell_op.run()
+            except Exception as error:  # skipcq PYL-W0703
+                logger.exception(error)
+
+            continue
+
+        if block["blockType"] == DBTCORE:
+            dbt_op = DbtCoreOperation.load(block["blockName"])
+            if os.path.exists(dbt_op.profiles_dir / "profiles.yml"):
+                os.unlink(dbt_op.profiles_dir / "profiles.yml")
+            try:
+                dbt_op.run()
+            except Exception as error:  # skipcq PYL-W0703
+                logger.exception(error)
